Making it smarter
Added modules for the storage of location data. As coordinates are added, location data will be stored in module. Older location data will be migrated as well. 

Also added the takeScreenshot module, with lets us "see" factor. The app can now "look" at the Case field(to be implemented where ever else needed), and ensure it is editable before attempting to make a change.